inibitore funziona ora

+++ src/master.c
@@ -17,7 +17,7 @@
 int msqid;               // ID della coda di messaggi
 pid_t attivatore_pid;    // PID del processo attivatore
 pid_t alimentazione_pid; // PID del processo alimentazione
-
+pid_t inibitore_pid;     // PID del processo inibitore
 /**
  * Stampa le statistiche della simulazione.
  * Usa un semaforo per garantire che nessun altro processo modifichi le statistiche durante la stampa.
@@ -30,7 +30,7 @@ void printStats()
         exit(EXIT_FAILURE);
     }
 
-    sem_stats = getSemaphoreSet();
+    sem_stats = getSemaphoreStatsSets();
 
     semLock(sem_stats); // Blocco del semaforo
 
@@ -59,7 +59,8 @@ void cleanup()
         kill(attivatore_pid, SIGINT);
     if (alimentazione_pid > 0)
         kill(alimentazione_pid, SIGINT);
-    if (alimentazione_pid < 0 && attivatore_pid < 0)
+    if(inibitore_pid > 0) kill(inibitore_pid, SIGINT);
+    if (alimentazione_pid < 0 && attivatore_pid < 0 && inibitore_pid < 0)
         killpg(*PID_MASTER, SIGTERM); // per gli atomi iniziali
     // timeout per evitare di rimanere bloccati indefinitamente
     time_t start_time = time(NULL);
@@ -221,6 +222,32 @@ void createAlimentazione()
         printf("[INFO] Master (PID: %d): Processo alimentazione creato con PID: %d\n", getpid(), pid);
     }
 }
+void createInibitore()
+{
+    pid_t pid = fork();
+
+    if (pid < 0)
+    { // Errore nella creazione del processo
+        perror("[ERROR] Master: Fork fallita durante la creazione di un Inibitore");
+        handle_meltdown(SIGUSR1);
+    }
+    else if (pid == 0)
+    {
+        printf("[INFO] Inibitore (PID: %d): Avvio processo inibitore\n", getpid());
+
+        // Esegue `inibitore`
+        if (execlp("./inibitore", "inibitore", NULL) == -1)
+        {
+            perror("[ERROR] Inibitore: execlp fallito durante l'esecuzione del processo inibitore");
+            exit(EXIT_FAILURE);
+        }
+    }
+    else
+    {
+        inibitore_pid = pid;
+        printf("[INFO] Master (PID: %d): Processo inibitore creato con PID: %d\n", getpid(), pid);
+    }
+}
 
 /**
  * Legge i parametri dal file di configurazione e li memorizza nella memoria condivisa.
@@ -315,28 +342,34 @@ int main()
     }
     else if (AtomGroup == 0)
     {
-        // Processo fantasma
-        printf("[INFO] Phantom (PID: %d): Inizio esecuzione\n", getpid());
         setpgid(0, 0); // Imposta il processo fantasma come leader del gruppo di processi
         pause(); // Metti il processo fantasma in pausa
         exit(EXIT_SUCCESS);
     }
 
     // Processo master
-    setpgid(0, AtomGroup); // Imposta il gruppo di processi del master al PID del processo fantasma
-
     alimentazione_pid = -1;
     attivatore_pid = -1;
+    inibitore_pid = -1;
     printf("[INFO] Master (PID: %d): Inizio esecuzione del programma principale il mio gruppo di processi è %d\n", getpid(), getpgrp());
 
-    sem_stats = getSemaphoreSet();
+    sem_stats = getSemaphoreStatsSets();
     sem_start = getSemaphoreStartset();
+    
     semLock(sem_stats); // Blocco del semaforo
     printf("[INFO] Master (PID: %d): Semaphore set initialized with ID: %d\n", getpid(), sem_stats);
+    //chiedi se si vuole attivare l'inibitore propontendo una scelta 0 spento 1 acceso
+    printf("Vuoi attivare l'inibitore? 0 per no 1 per si\n");
+    int inibitore;
+    if(scanf("%d", &inibitore) != 1)
+    {
+        perror("[ERROR] Master: E stata inserita una scelta non valida");
+        exit(EXIT_FAILURE);
+    }
 
     // Configura la memoria condivisa
     const char *shm_name = "/Parametres";
-    const size_t shm_size = 11 * sizeof(int);
+    const size_t shm_size = 12 * sizeof(int);
     void *shmParamsPtr = create_shared_memory(shm_name, shm_size);
 
     // Puntatori alle variabili nella memoria condivisa
@@ -351,7 +384,9 @@ int main()
     PID_MASTER = (int *)(shmParamsPtr + 8 * sizeof(int));
     isCleaning = (int *)(shmParamsPtr + 9 * sizeof(int));
     PID_GROUP_ATOMO = (int *)(shmParamsPtr + 10 * sizeof(int));
+    isinibitoreactive = (int *)(shmParamsPtr + 11 * sizeof(int));
     *PID_GROUP_ATOMO = AtomGroup;
+    *isinibitoreactive = inibitore;
     *isCleaning = 0;
     printf("[INFO] Master (PID: %d): Memoria condivisa mappata con successo. Inizio lettura del file di configurazione\n", getpid());
 
@@ -425,12 +460,19 @@ int main()
     createAlimentazione();
     waitForNInitMsg(msqid, 1);
     printf("---------------------------------------\n");
-
+    if(*isinibitoreactive == 1)
+    {
+        printf("[INFO] Master (PID: %d): Creazione del processo inibitore\n", getpid());
+        createInibitore();
+        waitForNInitMsg(msqid, 1);
+        printf("---------------------------------------\n");
+    }
     // Avvio della simulazione principale
     int termination = 0;
 
     printf("Attivatore PID: %d\n", attivatore_pid);
     printf("Alimentazione PID: %d\n", alimentazione_pid);
+    if(inibitore_pid > 0) printf("Inibitore PID: %d\n", inibitore_pid);
     semUnlock(sem_start); // Sblocco del semaforo
     printf("[INFO] Master( PID: %d): Inizio simulazione semaforo sbloccato\n", getpid());
